move file dialog converter to gin (#19709)
