+++ filenames.gni
@@ -466,7 +466,8 @@ filenames = {
     "shell/common/crash_reporter/win/crash_service_main.cc",
     "shell/common/crash_reporter/win/crash_service_main.h",
     "shell/common/gin_converters/callback_converter_gin_adapter.h",
-    "shell/common/gin_converters/file_dialog_converter_gin_adapter.h",
+    "shell/common/gin_converters/file_dialog_converter.cc",
+    "shell/common/gin_converters/file_dialog_converter.h",
     "shell/common/gin_converters/file_path_converter.h",
     "shell/common/gin_converters/gurl_converter.h",
     "shell/common/gin_converters/image_converter_gin_adapter.h",
@@ -497,7 +498,6 @@ filenames = {
     "shell/common/native_mate_converters/callback.h",
     "shell/common/native_mate_converters/content_converter.cc",
     "shell/common/native_mate_converters/content_converter.h",
-    "shell/common/native_mate_converters/file_dialog_converter.cc",
     "shell/common/native_mate_converters/file_dialog_converter.h",
     "shell/common/native_mate_converters/file_path_converter.h",
     "shell/common/native_mate_converters/gfx_converter.cc",
