+++ shell/common/gin_converters/file_dialog_converter.cc
@@ -1,21 +1,20 @@
-// Copyright (c) 2015 GitHub, Inc.
+// Copyright (c) 2019 GitHub, Inc.
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#include "shell/common/native_mate_converters/file_dialog_converter.h"
+#include "shell/common/gin_converters/file_dialog_converter.h"
 
-#include "native_mate/dictionary.h"
+#include "gin/dictionary.h"
 #include "shell/browser/api/atom_api_browser_window.h"
-#include "shell/browser/ui/file_dialog.h"
-#include "shell/common/native_mate_converters/file_path_converter.h"
-#include "shell/common/native_mate_converters/native_window_converter.h"
+#include "shell/common/gin_converters/file_path_converter.h"
+#include "shell/common/gin_converters/native_window_converter.h"
 
-namespace mate {
+namespace gin {
 
 bool Converter<file_dialog::Filter>::FromV8(v8::Isolate* isolate,
                                             v8::Local<v8::Value> val,
                                             file_dialog::Filter* out) {
-  mate::Dictionary dict;
+  gin::Dictionary dict(nullptr);
   if (!ConvertFromV8(isolate, val, &dict))
     return false;
   if (!dict.Get("name", &(out->first)))
@@ -28,19 +27,19 @@ bool Converter<file_dialog::Filter>::FromV8(v8::Isolate* isolate,
 v8::Local<v8::Value> Converter<file_dialog::Filter>::ToV8(
     v8::Isolate* isolate,
     const file_dialog::Filter& in) {
-  mate::Dictionary dict = mate::Dictionary::CreateEmpty(isolate);
+  gin::Dictionary dict = gin::Dictionary::CreateEmpty(isolate);
 
   dict.Set("name", in.first);
   dict.Set("extensions", in.second);
 
-  return dict.GetHandle();
+  return gin::ConvertToV8(isolate, dict);
 }
 
 bool Converter<file_dialog::DialogSettings>::FromV8(
     v8::Isolate* isolate,
     v8::Local<v8::Value> val,
     file_dialog::DialogSettings* out) {
-  mate::Dictionary dict;
+  gin::Dictionary dict(nullptr);
   if (!ConvertFromV8(isolate, val, &dict))
     return false;
   dict.Get("window", &(out->parent_window));
@@ -59,7 +58,7 @@ bool Converter<file_dialog::DialogSettings>::FromV8(
 v8::Local<v8::Value> Converter<file_dialog::DialogSettings>::ToV8(
     v8::Isolate* isolate,
     const file_dialog::DialogSettings& in) {
-  mate::Dictionary dict = mate::Dictionary::CreateEmpty(isolate);
+  gin::Dictionary dict = gin::Dictionary::CreateEmpty(isolate);
 
   dict.Set("window",
            electron::api::BrowserWindow::From(isolate, in.parent_window));
@@ -72,7 +71,7 @@ v8::Local<v8::Value> Converter<file_dialog::DialogSettings>::ToV8(
   dict.Set("showsTagField", in.shows_tag_field);
   dict.Set("securityScopedBookmarks", in.security_scoped_bookmarks);
 
-  return dict.GetHandle();
+  return gin::ConvertToV8(isolate, dict);
 }
 
-}  // namespace mate
+}  // namespace gin
