+++ shell/common/native_mate_converters/file_dialog_converter.h
@@ -6,26 +6,34 @@
 #define SHELL_COMMON_NATIVE_MATE_CONVERTERS_FILE_DIALOG_CONVERTER_H_
 
 #include "native_mate/converter.h"
-#include "shell/browser/ui/file_dialog.h"
+#include "shell/common/gin_converters/file_dialog_converter.h"
 
 namespace mate {
 
 template <>
 struct Converter<file_dialog::Filter> {
   static v8::Local<v8::Value> ToV8(v8::Isolate* isolate,
-                                   const file_dialog::Filter& in);
+                                   const file_dialog::Filter& in) {
+    return gin::ConvertToV8(isolate, in);
+  }
   static bool FromV8(v8::Isolate* isolate,
                      v8::Local<v8::Value> val,
-                     file_dialog::Filter* out);
+                     file_dialog::Filter* out) {
+    return gin::ConvertFromV8(isolate, val, out);
+  }
 };
 
 template <>
 struct Converter<file_dialog::DialogSettings> {
   static v8::Local<v8::Value> ToV8(v8::Isolate* isolate,
-                                   const file_dialog::DialogSettings& in);
+                                   const file_dialog::DialogSettings& in) {
+    return gin::ConvertToV8(isolate, in);
+  }
   static bool FromV8(v8::Isolate* isolate,
                      v8::Local<v8::Value> val,
-                     file_dialog::DialogSettings* out);
+                     file_dialog::DialogSettings* out) {
+    return gin::ConvertFromV8(isolate, val, out);
+  }
 };
 
 }  // namespace mate
